==> 1_training (1).ipynb <==
## Install Python dependencies

!pip install -r requirements.txt

Import the dependencies for the model training code:

import numpy as np
import pandas as pd
import datetime
from keras.models import Sequential
from keras.layers import Dense, Dropout, BatchNormalization, Activation
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.utils import class_weight
import tf2onnx
import onnx
import pickle
from pathlib import Path
import warnings
warnings.filterwarnings('ignore')

#Laoding the dataset
Data = pd.read_csv('caloric_exp.csv')
Data.head()

# Set the input (X) and output (Y) data.

X = Data.drop(columns = ['user_id','date','recommended_calorie_intake_in_kcal_per_day'])
y = Data['recommended_calorie_intake_in_kcal_per_day']


# Split the data into training and testing sets so you have something to test the trained model with.
X_train, X_test, y_train, y_test = train_test_split(X,y, test_size = 0.2)
X_train, X_val, y_train, y_val = train_test_split(X_train,y_train, test_size = 0.2)

Path("artifact").mkdir(parents=True, exist_ok=True)
with open("artifact/test_data.pkl", "wb") as handle:
    pickle.dump((X_test, y_test), handle)

#fetching the cols:
num_cols = X_train.select_dtypes(include='number').columns.to_list()
cat_cols = X_train.select_dtypes(exclude='number').columns.to_list()

#encoding the categorical_variables:
from sklearn.preprocessing import OneHotEncoder
ohe = OneHotEncoder(drop='first',handle_unknown='ignore',sparse_output=False)

#importing the scaler
scaler = StandardScaler()

from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

ct = ColumnTransformer(
    transformers = [('Encoder',OneHotEncoder(drop='first',handle_unknown='ignore',sparse_output=False),cat_cols),
                    ('Scaler',StandardScaler(),num_cols)],
                    remainder='passthrough')


ct_01= ct.fit(X_train)

with open("artifact/column_transformer.pkl", "wb") as handle:
    pickle.dump(ct_01, handle)

## Build the model

The model is a simple, fully-connected, deep neural network, containing three hidden layers and one output layer.

X_train_new = ct.fit_transform(X_train)

model = Sequential()
model.add(Dense(32, activation = 'relu', input_dim = X_train_new.shape[1]))
# model.add(Dropout(0.2))
# model.add(Dense(32))
# model.add(BatchNormalization())
# model.add(Activation('relu'))
# model.add(Dropout(0.2))
# model.add(Dense(32))
# model.add(BatchNormalization())
# model.add(Activation('relu'))
# model.add(Dropout(0.2))
model.add(Dense(1, activation = 'linear'))
model.compile(optimizer='adam',loss='mean_squared_error',metrics=['r2_score'])
model.summary()

## Train the model

Training a model is often the most time-consuming part of the machine learning process.  Large models can take multiple GPUs for days.  Expect the training on CPU for this very simple model to take a minute or more.

# Train the model and get performance
import os

epochs = 200
history = model.fit(X_train_new, y_train.values, epochs=epochs, \
                    validation_data=(ct.transform(X_val),y_val), \
                    verbose = True)
print("Training of model is complete")

## Save the model file

os.makedirs("models/caloric/1", exist_ok=True)

saved_model_path = "models/caloric/1/my_model"
model.export(saved_model_path)

X_test.dropna().sample(5).to_csv('X_test.csv',index = False)

X_test.columns

X_test.iloc[0,:].values

!pip install tf2onnx

#!mo --saved_model_dir models/caloric/1/my_model --output_dir models/caloric/1 --framework tf

!python -m tf2onnx.convert --saved-model models/caloric/1/my_model --output model.onnx
