==> 3_rest_requests_multi_model (1) (1).ipynb <==
# REST Inference

## Setup

Verify that following variable settings match your deployed model's resource name and rest URL. The following code assumes that the kube service is in the same namespace, but you could refer to it in full with the namespace, for example: `http://modelmesh-serving.project-name.svc.cluster.local:8008/v2/models/fraud/infer`

infer_url = "https://caloricexp-neeraj-jain-dev.apps.sandbox-m3.1530.p1.openshiftapps.com/v2/models/caloricexp/infer"

## Request Function

Build and submit the REST request. 

Note: You submit the data in the same format that you used for an ONNX inference.

import requests

def rest_request(data):
    json_data = {
        "inputs": [
            {
                "name": "keras_tensor",
                "shape": [1, 35],
                "datatype": "FP32",
                "data": data
            }
        ]
    }

    response = requests.post(infer_url, json=json_data)
    response_dict = response.json()
    return response_dict['outputs'][0]['data']

#Load the scaler
import pickle
with open('artifact/column_transformer.pkl', 'rb') as handle:
    ct = pickle.load(handle)

import pandas as pd

# Define your input data as a list
data = pd.read_csv('X_test.csv')
data

# Transform the DataFrame using the column transformer
transformed_data = ct.transform(data.sample(1))
transformed_data

transformed_data.shape

prediction = rest_request(transformed_data.tolist())
prediction
